DW-Plant-e/requirements.txt
# Run:
# pip install -r requirements.txt

# This is an implicit value, here for clarity
# --index-url specifies the url to use when downloading packages
--index-url https://pypi.python.org/simple/

# Below are the packages needed
pyrebase4==4.3.0
serial==0.0.97
smbus==1.1
picamera==1.13

# This looks at the github account, setup.py file and downloads the modules from there
-e git+git://github.com/LoJunKai/DW-Plant-e.git@master#egg=plante  # egg=<package_name> (same as the name in setup.py)


# Source: https://caremad.io/posts/2013/07/setup-vs-requirement/
